# packages/api-types/src/index.ts
export {
  ACCOUNT_DISABLED_ISSUED_TOKEN,
  ACCOUNT_DISABLED_ON_LOGIN,
  API_ERROR_CODES,
  API_ERROR_DEFS,
  CONFLICT_BINDING_MESSAGE,
  FORBIDDEN_READ_MESSAGE,
  type ApiErrorCode,
} from './error-codes';
export { ApiError, mapErrorBody, type ApiErrorBody } from './errors';
export {
  API_V1_PREFIX,
  BINDING_KINDS,
  COURSE_SUMMARY_MAX_LENGTH,
  P0_PATHS,
  P1_HOME_PATHS,
  POSTER_TEMPLATE_KEYS,
  PUBLIC_FEATURED_ARTWORK_FIELDS,
  ROLES,
  STUDENT_STATUSES,
  USER_STATUSES,
  WATERMARK_POSITIONS,
  type Account,
  type AdminCarousel,
  type AdminCourse,
  type AdminFeaturedArtwork,
  type Artwork,
  type ArtworkComment,
  type ArtworkCursorPage,
  type ArtworkCursorQuery,
  type AuthUser,
  type Binding,
  type BindingKind,
  type BindingQuery,
  type Brand,
  type CreateAccountRequest,
  type CreateArtworkFields,
  type CreateBindingRequest,
  type CreateCarouselRequest,
  type CreateCommentRequest,
  type CreateCourseRequest,
  type CreateFeaturedArtworkRequest,
  type CreateStudentRequest,
  type GeneratePosterRequest,
  type HomeAdminList,
  type LoginRequest,
  type LoginResponse,
  type LogoutRequest,
  type LogoutResponse,
  type P0ErrorCode,
  type PosterDownload,
  type PosterPreview,
  type PosterTemplateKey,
  type PublicCarousel,
  type PublicCourse,
  type PublicFeaturedArtwork,
  type PublicHome,
  type PublicHomeBrand,
  type RefreshRequest,
  type RefreshResponse,
  type Role,
  type Student,
  type StudentStatus,
  type TokenPair,
  type UpdateAccountRequest,
  type UpdateAccountStatusRequest,
  type UpdateBrandRequest,
  type UpdateCarouselRequest,
  type UpdateCourseRequest,
  type UpdateFeaturedArtworkRequest,
  type UpdateStudentRequest,
  type UserStatus,
  type WatermarkPosition,
} from './dto';
export {
  EXAMPLE_ACCOUNT,
  EXAMPLE_ACCOUNT_DISABLED,
  EXAMPLE_ACCOUNT_DISABLED_ISSUED_TOKEN,
  EXAMPLE_ARTWORK,
  EXAMPLE_ARTWORK_PAGE,
  EXAMPLE_AUTH_USER,
  EXAMPLE_BINDING,
  EXAMPLE_BRAND,
  EXAMPLE_CONFLICT_BINDING,
  EXAMPLE_CONFLICT_STUDENT_HAS_ARTWORK,
  EXAMPLE_FORBIDDEN_READ,
  EXAMPLE_LOGIN_RESPONSE,
  EXAMPLE_LOGO_NOT_CONFIGURED,
  EXAMPLE_POSTER_DOWNLOAD,
  EXAMPLE_POSTER_DOWNLOAD_URL,
  EXAMPLE_POSTER_PREVIEW,
  EXAMPLE_POSTER_PREVIEW_URL,
  EXAMPLE_STUDENT,
  EXAMPLE_PUBLIC_FEATURED_ARTWORK,
  EXAMPLE_PUBLIC_HOME,
} from './examples';